refactor: extract ANSI, format, and UI shared components from AppView.tsx
Issue #40: Dashboard Component Refactor (Mandatory)

- Extracted ANSI/Markdown utilities to src/lib/ansi.ts

- Extracted formatting helpers to src/lib/format.ts

- Extracted shared UI primitives (StatusDot, PhaseBadge, ConnectionIndicator, ElapsedTimer) to src/components/shared/
