+++ aloop/cli/dashboard/src/components/shared/StatusDot.tsx
@@ -1,0 +1,33 @@
+import React from 'react';
+import { Tooltip, TooltipContent, TooltipTrigger } from '@/components/ui/tooltip';
+
+const STATUS_DOT_CONFIG: Record<string, { color: string; label: string }> = {
+  running: { color: 'bg-green-500', label: 'Running' },
+  stopped: { color: 'bg-muted-foreground/50', label: 'Stopped' },
+  exited: { color: 'bg-muted-foreground/50', label: 'Exited' },
+  unhealthy: { color: 'bg-red-500', label: 'Unhealthy' },
+  error: { color: 'bg-red-500', label: 'Error' },
+  stuck: { color: 'bg-orange-500', label: 'Stuck' },
+  unknown: { color: 'bg-muted-foreground/30', label: 'Unknown' },
+};
+
+export function StatusDot({ status, className = '' }: { status: string; className?: string }) {
+  const config = STATUS_DOT_CONFIG[status] ?? STATUS_DOT_CONFIG.unknown;
+  const label = config.label;
+
+  const dot = status === 'running' ? (
+    <span className={`relative flex h-2.5 w-2.5 shrink-0 ${className}`}>
+      <span className={`absolute inline-flex h-full w-full animate-pulse-dot rounded-full ${config.color}/70`} />
+      <span className={`relative inline-flex h-2.5 w-2.5 rounded-full ${config.color}`} />
+    </span>
+  ) : (
+    <span className={`inline-flex h-2.5 w-2.5 shrink-0 rounded-full ${config.color} ${className}`} />
+  );
+
+  return (
+    <Tooltip>
+      <TooltipTrigger asChild><span className="inline-flex">{dot}</span></TooltipTrigger>
+      <TooltipContent><p>{label}</p></TooltipContent>
+    </Tooltip>
+  );
+}
